feat: move dod settlement contract api to enterprise build
