+++ cmd/client/commands/dod_settlement.go
@@ -1,3 +1,5 @@
+// +build testnet
+
 package commands
 
 import (
