+++ cmd/client/commands/dod_settlement_terminate_response.go
@@ -1,3 +1,5 @@
+// +build testnet
+
 package commands
 
 import (
@@ -5,14 +7,13 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/qlcchain/go-qlc/rpc/api"
-
 	"github.com/abiosoft/ishell"
 	rpc "github.com/qlcchain/jsonrpc2"
 
 	"github.com/qlcchain/go-qlc/cmd/util"
 	"github.com/qlcchain/go-qlc/common/types"
 	cutil "github.com/qlcchain/go-qlc/common/util"
+	"github.com/qlcchain/go-qlc/rpc/api"
 	"github.com/qlcchain/go-qlc/vm/contract/abi"
 )
