+++ cmd/client/commands/helper.go
@@ -1,0 +1,21 @@
+// +build !testnet
+
+package commands
+
+func addcommands() {
+	version()
+	addAccountCmd()
+	addDebugCmd()
+	addLedgerCmd()
+	addMinerCmd()
+	addWalletCmd()
+	addMintageCmd()
+	addPovCmd()
+	addPledgeCmd()
+	addRepCmd()
+	addTxCmd()
+	addPKDCmd()
+	addPermissionCmd()
+	addPrivacyCmd()
+	addPtmKeyCmd()
+}
