+++ cmd/client/commands/helper_testnet.go
@@ -1,0 +1,22 @@
+// +build testnet
+
+package commands
+
+func addcommands() {
+	version()
+	addAccountCmd()
+	addDebugCmd()
+	addLedgerCmd()
+	addMinerCmd()
+	addWalletCmd()
+	addMintageCmd()
+	addPovCmd()
+	addPledgeCmd()
+	addRepCmd()
+	addTxCmd()
+	addPKDCmd()
+	addPermissionCmd()
+	addPrivacyCmd()
+	addPtmKeyCmd()
+	addDoDSettlementCmd()
+}
