+++ config/config_v8.go
@@ -8,6 +8,7 @@ func DefaultConfigV8(dir string) (*ConfigV8, error) {
 	var cfg ConfigV8
 	cfg7, _ := DefaultConfigV7(dir)
 	cfg.ConfigV7 = *cfg7
+	cfg.RPC.PublicModules = defaultModules()
 	cfg.RPC.GRPCConfig = defaultGRPCConfig()
 	return &cfg, nil
 }
