+++ rpc/api/dod_settlement_test.go
@@ -1,3 +1,5 @@
+// +build testnet
+
 package api
 
 import (
