+++ rpc/grpc/apis/dod_settlement.go
@@ -1,0 +1,5 @@
+// +build testnet
+
+package apis
+
+//TODO: implement dod settlement apis
